+++ sdk/rust/src/allocator/mod.rs
@@ -1,0 +1,30 @@
+use std::alloc::{Alloc, Global, Layout};
+
+/// Allocates a memory region of given size and returns its address. Actually is
+/// just a wrapper for [`GlobalAlloc::alloc`].
+///
+/// # Safety
+///
+/// See [`GlobalAlloc::alloc`].
+///
+/// [`GlobalAlloc::alloc`]: https://doc.rust-lang.org/core/alloc/trait.GlobalAlloc.html#tymethod.alloc
+///
+pub unsafe fn alloc(size: NonZeroUsize) -> MemResult<NonNull<u8>> {
+    let layout: Layout = Layout::from_size_align(size.get(), mem::align_of::<u8>())?;
+    Global.alloc(layout).map_err(Into::into)
+}
+
+/// Deallocates a memory region for current pointer and size. Actually is
+/// just a wrapper for [`GlobalAlloc::dealloc`].
+///
+/// # Safety
+///
+/// See [`GlobalAlloc::dealloc`].
+///
+/// [`GlobalAlloc::dealloc`]: https://doc.rust-lang.org/core/alloc/trait.GlobalAlloc.html#tymethod.dealloc
+///
+pub unsafe fn dealloc(ptr: NonNull<u8>, size: NonZeroUsize) -> MemResult<()> {
+    let layout = Layout::from_size_align(size.get(), mem::align_of::<u8>())?;
+    Global.dealloc(ptr, layout);
+    Ok(())
+}
